requirements.txt
# Core dependencies
anthropic>=0.39.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# GitHub integration
PyGithub>=2.1.1
gitpython>=3.1.40

# CLI and utilities
click>=8.1.7
rich>=13.7.0
python-dateutil>=2.8.2

# State management
redis>=5.0.1

# HTTP client
httpx>=0.25.2

# Logging and monitoring
structlog>=24.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code quality
black>=23.12.1
ruff>=0.1.9
mypy>=1.7.1

# Optional: async support
aiohttp>=3.9.1
